addedd helper function to handle board permissions and ownership.

+++ src/app/dashboard/projects/actions.ts
@@ -28,6 +28,7 @@ import {
 import { CardUpdate } from '@/use-cases/types';
 import { deleteBoard } from '@/data-access/projects';
 import { redirect } from 'next/navigation';
+import { checkBoardOwnership } from '@/util/auth-projects';
 //import type { CardUpdate } from '@/db/schema/projects';
 
 const projectSchema = z.object({
@@ -70,6 +71,8 @@ export const updateBoardAction = authenticatedAction
       key: `${user.id}-update-project`
     });
 
+    await checkBoardOwnership(boardId, user.id);
+
     await projectsDb.updateBoard(boardId, { name, description }, user);
     revalidatePath(`/dashboard/projects/${boardId}`);
   });
@@ -78,6 +81,7 @@ export const deleteBoardAction = authenticatedAction
   .createServerAction()
   .input(z.object({ boardId: z.number() }))
   .handler(async ({ input: { boardId }, ctx: { user } }) => {
+    await checkBoardOwnership(boardId, user.id);
     await deleteBoard(boardId, user);
     revalidatePath('/dashboard/projects');
     redirect('/dashboard/projects');
